+++ src/entities/reviews/index.ts
@@ -4,5 +4,5 @@ export {default as ReviewList} from './ui/ReviewList';
 export {default as ReviewArticle} from './ui/ReviewArticle';
 export {default as ReviewArticleLoading} from './ui/ReviewArticleLoading';
 export {default as useCategoryReviews} from './model/useCategoryReviews';
-export {default as useSearchReviewsWithKeyword} from './model/useSearchReviewsWithQuery';
+export {default as useKeywordReviews} from './model/useKeywordReviews';
 export {default as ReviewsLoading} from './ui/ReviewsLoading';
