+++ src/features/reviews/keyword-search/ui/ReviewWithPagination.tsx
@@ -3,7 +3,7 @@
 import {useParams, useSearchParams} from 'next/navigation';
 import Pagination from '@/widgets/pagination';
 import {SortKey} from '@/features/reviews/sorting';
-import {ReviewArticle, useSearchReviewsWithKeyword} from '@/entities/reviews';
+import {ReviewArticle, useKeywordReviews} from '@/entities/reviews';
 
 type Props = {
   sort: SortKey;
@@ -14,7 +14,7 @@ export default function ReviewWithPagination({sort}: Props) {
   const searchParams = useSearchParams();
 
   const currentPage = Number(searchParams.get('page')) || 1;
-  const {results, total_pages} = useSearchReviewsWithKeyword(keyword, currentPage, sort);
+  const {results, total_pages} = useKeywordReviews(keyword, currentPage, sort);
 
   return (
     <ul className="flex flex-col mb-6">
